@@ -1,7 +1,7 @@
 package com.svmdev.appmovies;
 
 import android.content.Intent;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 
 import com.svmdev.appmovies.view.MainActivity;
